Played with deploy

+++ resources/ActionScripts/LoadImageData.ps1
@@ -61,11 +61,13 @@ Write-Host (
     "Data Loaded and Trained in $Duration")
 
 ##Remove Run Once
+if ($isDeploy -eq "No")
+{
 Remove-Item "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\StartUp\RunOnce.cmd"
+}
 
 ##Copy Url to Start Menu
 Copy-Item "C:\Solutions\ImageSimilarity\Resources\ActionScripts\SolutionHelp.url" "C:\ProgramData\Microsoft\Windows\Start Menu\Programs\StartUp\"
 
-##Launch HelpURL 
-Start-Process https://microsoft.github.io/ml-server-image-similarity/
+Read-Host "Press the Enter Key To Continue"
 }
